+++ packages/gamut-labs/src/AppHeaderMobile/AppHeaderSubMenuMobile/index.tsx
@@ -9,13 +9,28 @@ import { AppHeaderCatalogSection } from '../../AppHeader/AppHeaderElements/AppHe
 import { AppHeaderDropdownProps } from '../../AppHeader/AppHeaderElements/AppHeaderDropdown';
 import { AppHeaderLinkSections } from '../../AppHeader/AppHeaderElements/AppHeaderLinkSections';
 import { AppHeaderListItem } from '../../AppHeader/AppHeaderElements/AppHeaderListItem';
+import { AppHeaderResourceDropdownProps } from '../../AppHeader/AppHeaderElements/AppHeaderResourcesDropdown';
+import { AppHeaderResourcesSection } from '../../AppHeader/AppHeaderElements/AppHeaderResourcesSection';
+import {
+  AppHeaderCatalogDropdownItem,
+  AppHeaderClickHandler,
+  AppHeaderDropdownItem,
+  AppHeaderResourcesDropdownItem,
+} from '../../AppHeader/AppHeaderElements/types';
 
 export type AppHeaderSubMenuMobileProps = (
   | AppHeaderDropdownProps
   | AppHeaderCatalogDropdownProps
+  | AppHeaderResourceDropdownProps
 ) & {
-  handleClose: () => void;
+  handleCloseSubMenu: () => void;
+  handleCloseMainMenu: () => void;
 };
+
+type AppHeaderSectionItem =
+  | AppHeaderDropdownItem
+  | AppHeaderCatalogDropdownItem
+  | AppHeaderResourcesDropdownItem;
 
 const StyledAnchor = styled(Anchor)(
   css({
@@ -27,15 +42,44 @@ const StyledAnchor = styled(Anchor)(
   })
 );
 
+const renderHeaderSection = (
+  item: AppHeaderSectionItem,
+  action: AppHeaderClickHandler,
+  handleCloseMainMenu: () => void
+) => {
+  switch (item.type) {
+    case 'catalog-dropdown':
+      return (
+        <AppHeaderCatalogSection
+          action={action}
+          item={item}
+          handleClose={handleCloseMainMenu}
+        />
+      );
+    case 'experimental-resources-dropdown':
+      return (
+        <AppHeaderResourcesSection
+          action={action}
+          handleClose={handleCloseMainMenu}
+        />
+      );
+    default:
+      return (
+        <AppHeaderLinkSections action={action} item={item} showIcon mobile />
+      );
+  }
+};
+
 export const AppHeaderSubMenuMobile: React.FC<AppHeaderSubMenuMobileProps> = ({
   action,
-  handleClose,
+  handleCloseSubMenu,
+  handleCloseMainMenu,
   item,
 }) => {
   return (
     <AppHeaderListItem aria-labelledby={`${item.text} menu`}>
       <StyledAnchor
-        onClick={handleClose}
+        onClick={handleCloseSubMenu}
         variant="interface"
         as="button"
         ml={{ _: 16, xs: 32, sm: 64, md: 48 }}
@@ -53,11 +97,7 @@ export const AppHeaderSubMenuMobile: React.FC<AppHeaderSubMenuMobileProps> = ({
       >
         {item.type === 'profile-dropdown' ? item.userDisplayName : item.text}
       </Text>
-      {item.type === 'catalog-dropdown' ? (
-        <AppHeaderCatalogSection action={action} item={item} />
-      ) : (
-        <AppHeaderLinkSections action={action} item={item} showIcon mobile />
-      )}
+      {renderHeaderSection(item, action, handleCloseMainMenu)}
     </AppHeaderListItem>
   );
 };